+++ pom.xml
@@ -54,8 +54,7 @@
   <org.apache.commons.validator.version>1.6</org.apache.commons.validator.version>
   <org.apache.commons.io.version>2.6</org.apache.commons.io.version>
   <org.apache.httpcomponents.version>4.5.11</org.apache.httpcomponents.version>
-  <io.fabric8.openshift.client.version>4.6.4</io.fabric8.openshift.client.version>
-  <io.kubernetes.client.java.version>5.0.0</io.kubernetes.client.java.version>
+  <io.fabric8.client.version>4.6.4</io.fabric8.client.version>
   <io.vertx.web.version>3.9.2</io.vertx.web.version>
   <org.slf4j.version>1.7.30</org.slf4j.version>
   <com.google.code.gson.version>2.8.6</com.google.code.gson.version>
@@ -81,12 +80,12 @@
   <dependency>
     <groupId>io.fabric8</groupId>
     <artifactId>openshift-client</artifactId>
-    <version>${io.fabric8.openshift.client.version}</version>
+    <version>${io.fabric8.client.version}</version>
   </dependency>
   <dependency>
-    <groupId>io.kubernetes</groupId>
-    <artifactId>client-java</artifactId>
-    <version>${io.kubernetes.client.java.version}</version>
+    <groupId>io.fabric8</groupId>
+    <artifactId>kubernetes-client</artifactId>
+    <version>${io.fabric8.client.version}</version>
   </dependency>
   <dependency>
     <groupId>com.google.dagger</groupId>
