+++ src/main/java/com/redhat/rhjmc/containerjfr/platform/internal/KubeApiPlatformClient.java
@@ -41,6 +41,7 @@
  */
 package com.redhat.rhjmc.containerjfr.platform.internal;
 
+import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
 import java.util.Objects;
@@ -48,75 +49,79 @@ import java.util.stream.Collectors;
 
 import com.redhat.rhjmc.containerjfr.core.log.Logger;
 import com.redhat.rhjmc.containerjfr.core.net.JFRConnectionToolkit;
-import com.redhat.rhjmc.containerjfr.net.NetworkResolver;
 import com.redhat.rhjmc.containerjfr.platform.PlatformClient;
 import com.redhat.rhjmc.containerjfr.platform.ServiceRef;
 
 import dagger.Lazy;
-import io.kubernetes.client.ApiException;
-import io.kubernetes.client.apis.CoreV1Api;
-import io.kubernetes.client.models.V1Service;
+import io.fabric8.kubernetes.api.model.EndpointPort;
+import io.fabric8.kubernetes.api.model.EndpointSubset;
+import io.fabric8.kubernetes.api.model.Endpoints;
+import io.fabric8.kubernetes.client.KubernetesClient;
 
 class KubeApiPlatformClient implements PlatformClient {
 
-    private final CoreV1Api api;
-    private final String namespace;
+    private final KubernetesClient k8sClient;
     private final Lazy<JFRConnectionToolkit> connectionToolkit;
-    private final NetworkResolver resolver;
     private final Logger logger;
+    private final String namespace;
 
     KubeApiPlatformClient(
-            CoreV1Api api,
             String namespace,
+            KubernetesClient k8sClient,
             Lazy<JFRConnectionToolkit> connectionToolkit,
-            NetworkResolver resolver,
             Logger logger) {
-        this.api = api;
         this.namespace = namespace;
+        this.k8sClient = k8sClient;
         this.connectionToolkit = connectionToolkit;
-        this.resolver = resolver;
         this.logger = logger;
     }
 
     @Override
     public List<ServiceRef> listDiscoverableServices() {
         try {
-            return api
-                    .listNamespacedService(
-                            namespace, null, null, null, null, null, null, null, null, null)
-                    .getItems().stream()
-                    .map(V1Service::getSpec)
-                    .peek(spec -> logger.trace("Service spec: " + spec.toString()))
-                    .filter(s -> s.getPorts() != null)
-                    .flatMap(
-                            s ->
-                                    s.getPorts().stream()
-                                            .map(
-                                                    p -> {
-                                                        try {
-                                                            return new ServiceRef(
-                                                                    connectionToolkit.get(),
-                                                                    s.getClusterIP(),
-                                                                    p.getPort(),
-                                                                    resolver
-                                                                            .resolveCanonicalHostName(
-                                                                                    s
-                                                                                            .getClusterIP()));
-                                                        } catch (Exception e) {
-                                                            logger.warn(e);
-                                                            return null;
-                                                        }
-                                                    }))
-                    .parallel()
-                    .filter(Objects::nonNull)
+            return k8sClient.endpoints().inNamespace(namespace).list().getItems().stream()
+                    .flatMap(endpoints -> getServiceRefs(endpoints).stream())
                     .collect(Collectors.toList());
-        } catch (ApiException e) {
-            logger.warn(e.getMessage());
-            logger.warn(e.getResponseBody());
-            return Collections.emptyList();
         } catch (Exception e) {
             logger.warn(e);
             return Collections.emptyList();
         }
+    }
+
+    private boolean isCompatiblePort(EndpointPort port) {
+        return "jfr-jmx".equals(port.getName()) || 9091 == port.getPort();
+    }
+
+    private List<ServiceRef> getServiceRefs(Endpoints endpoints) {
+        List<ServiceRef> refs = new ArrayList<>();
+        endpoints.getSubsets().stream()
+                .forEach(
+                        subset ->
+                                subset.getPorts().stream()
+                                        .filter(this::isCompatiblePort)
+                                        .forEach(
+                                                port ->
+                                                        refs.addAll(
+                                                                createServiceRefs(subset, port))));
+        return refs;
+    }
+
+    private List<ServiceRef> createServiceRefs(EndpointSubset subset, EndpointPort port) {
+        return subset.getAddresses().stream()
+                .map(
+                        addr -> {
+                            try {
+                                return new ServiceRef(
+                                        connectionToolkit.get(),
+                                        addr.getIp(),
+                                        port.getPort(),
+                                        addr.getTargetRef().getName());
+                            } catch (Exception e) {
+                                logger.warn(e);
+                                return null;
+                            }
+                        })
+                .filter(Objects::nonNull)
+                .collect(Collectors.toList());
     }
 }
